% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitland.R
\name{get_transit_route_stop_patterns}
\alias{get_transit_route_stop_patterns}
\title{Get transit route stop patterns}
\usage{
get_transit_route_stop_patterns(params, get_all = FALSE)
}
\arguments{
\item{parms}{list of query parameters. Valid parameters are
bbox
location}

\item{paginate}{TRUE or FALSE, will try to retrieve all results if TRUE. If per_page is set it will iterate
through pages}
}
\description{
Get transit route stop patterns
}
\examples{
get_transit_route_stop_patterns(list(bbox=c(-122.4183,37.7758,-122.4120,37.7858)))
}
