% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitland.R
\name{get_transit_schedule_stops}
\alias{get_transit_schedule_stops}
\title{Get schedule stop pairs}
\usage{
get_transit_schedule_stops(params, get_all = TRUE)
}
\arguments{
\item{parms}{list of query parameters. Valid parameters are
bbox
location}

\item{paginate}{TRUE or FALSE, will try to retrieve all results if TRUE. If per_page is set it will iterate
through pages}
}
\description{
Get schedule stop pairs
}
\examples{
get_transit_schedule_stops(list(bbox=c(-122.4183,37.7758,-122.4120,37.7858)))
}
