% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitland.R
\name{get_transit_stops}
\alias{get_transit_stops}
\title{Get stops data}
\usage{
get_transit_stops(params, get_all = FALSE)
}
\description{
Get stops data
}
\examples{
get_transit_stops(list(bbox=c(-122.4183,37.7758,-122.4120,37.7858)))
}
